fix: publishing move events to other nvim plugins
It would silently crash when trying to publish move events to other nvim
plugins. I don't think this is a huge issue since the feature does not
seem to be widely used.

It might be that it never actually worked for move events, but did work
for rename and bulk events.

For a description of this event publishing feature, see
https://github.com/mikavilpas/yazi.nvim/pull/495