feat: open yazi tabs for the files in the quickfix list
Issue
=====

The quickfix list is a versatile tool in neovim, and it can be used for
many different workflows.

yazi.nvim can send files from yazi to the quickfix list, but when this
is done, yazi is closed and the context is lost. When doing project wide
work both in yazi and neovim, it's difficult because of this.

Solution
========

Allow opening yazi tabs for the files in the quickfix list. This only
works if the user has enabled the `open_multiple_tabs` setting.
Otherwise the first quickfix file is selected.

When the quickfix list is open and focused, opening yazi will now open
(probably) 9 of the first files as yazi tabs. The limitation is in place
because yazi can open up to 9 tabs at a time.