fix: don't restore the previous buffer when yazi is closed
Sometimes when using vertical splits, I noticed that when yazi is
closed, one of the splits switches to a different buffer. I was unable
to reproduce this issue consistently, but I have had this happen every
once in a while for the last two weeks.

It seems like I added some logic to on_yazi_exited that restores the
previous buffer, but I don't think that's necessary - it does not seem
to be doing anything.

I'm hoping this will resolve the issue and not cause any (untested)
features to break, but I think it should be quite safe.