test: force test yazi to exit with `i` to test it's not sent
In the pr below, I tried to fix a bug where yazi.nvim would send `i` to
yazi.
https://github.com/mikavilpas/yazi.nvim/pull/876

This commit adds a yazi keymap for `i` to test that it's not sent.
Sending it would cause yazi to exit, which would fail any test that
expects yazi to be running.