feat: allow a yazi keymap to cycle through open buffers (#894)
Issue
=====

By default, the `<tab>` key in yazi.nvim is used to cycle through the
open buffers. This is done in a way that prevents using the `<tab>` key
for any yazi actions.

Most of the time this does not cause issues, but in some cases the
`<tab>` key is very important. For example, if a subshell is opened in
yazi using this mapping:

```toml
[manager]
keymap = {
  { on = [":"], run = "shell 'fish' --block --confirm" }
}
```
(https://github.com/mikavilpas/dotfiles/blob/96f734f27f449783b0a1884623e27fc567849d50/.config/yazi/keymap.toml?plain=1#L116)

In this case, the subshell has no access to the `<tab>` key. For example,
this disables tab completion in the user's shell.

Solution
========

Allow users to work around this by manually defining a yazi keymap that
sends an event to yazi.nvim. Yazi.nvim will then react to the event and
execute the cycle_open_buffers action.

To enable this mapping, look at how the integration tests define the
keymapping in
https://github.com/mikavilpas/yazi.nvim/blob/main/integration-tests/test-environment/.config/yazi/keymap.toml?plain=1