fix: not being able to cycle buffers with `tab` after `:Yazi cwd` (#922)
Issue
=====

After `:Yazi cwd`, the `tab` keybinding (for `cycle_open_buffers`) does
nothing.

The old implementation of cycling buffers depended on the initial file
that was opened - it was incorrectly presumed that the buffer is one of
the currently open buffers. This is not true when opening a directory.

Solution
========

Pick the first window as the initial target (the buffer to switch focus
to) when `:Yazi cwd` is used.

Closes https://github.com/mikavilpas/yazi.nvim/issues/914