feat(snacks): add copy relative path action to snacks.nvim picker (#938)
yazi.nvim has had the ability to copy the relative path to files for a
while now (default `<c-y>`). This commit adds the ability to do the same in
the snacks.nvim picker. It's not enabled by default.

When the snacks.nvim picker is open, you can press your keybinding for
this, and it will copy the relative paths to the selected file(s) to the
clipboard register.

To learn how to use this, you can read the documentation in the readme
file.