feat: allow emitting arbitrary yazi commands after yazi is ready (#946)
When Yazi is running, it supports receiving instructions to execute yazi
commands. These can be sent using `ya emit-to`:
- https://yazi-rs.github.io/docs/dds#ya-emit

Provide two things to take advantage of this:
- a new `on_yazi_ready` hook that is called when yazi is ready to
  receive commands
- a new `emit_to_yazi` method on the YaziProcessApi that allows
  sending commands to yazi.

This allows starting yazi in various interesting states, such as
immediately entering find mode.

More detailed documentation can be found in the features section of the
readme.

Closes https://github.com/mikavilpas/yazi.nvim/issues/944