refactor: remove fallbacks for missing winborder option
yazi.nvim supports the `winborder` option added in Neovim 0.11. Back
when this feature was added, it was not available in stable Neovim, so I
added protections to prevent errors when the option was not available.

Now that Neovim 0.11 is the minimum supported version, we can remove
these fallbacks and simplify the code.