refactor: make snacks.nvim optional by bundling snacks.bufdelete (#1019)
**Issue:**

For a long time, snacks.nvim had been a dependency of yazi.nvim.
Originally I had big plans for using it extensively, but they have not
materialized (yet).

Snacks is still a dependency by default, although only a very small part
of it is used: the `snacks.bufdelete` module, which is used to delete
buffers while preserving the window layout.

Previously, `future_features.process_events_live` has been set to `true`
by default, but because the built-in way of closing buffers was found to
disturb the window layout, it was recommended to use `snacks.bufdelete`
instead.

**Solution:**

Bundle a copy of `snacks.bufdelete` in yazi.nvim, so that snacks.nvim is
not required anymore.

Turn this on by default, but allow users to disable it by setting
`config.integrations.bufdelete_implementation = "builtin"` and
`config.future_features.process_events_live2 = false`. If this happens
to you, please open an issue, as this will be the default in the future.