test: use internal way of polling for yazi to be ready (#1020)
**Issue:**

Previously, the tests were waiting for a specific text in yazi's footer
to be visible to determine if yazi was ready. The terminal can sometimes
shift out of view in CI, so this is not a reliable way.

**Solution:**

Use the internal way of polling for yazi to be ready
(`assertYaziIsReady`), which was added some time ago but is not used in
all of the tests yet.