test: correctly call `vim.lsp.enable` to enable LSP in tests
**Issue:**

Previously, the `vim.lsp.enable("lua-language-server")` call was made in
the test environment's `init.lua` file, but it was using the wrong
server name.

The tests still worked because the `mason-lspconfig` plugin was
automatically enabling the `lua_ls` as it's installed on my system.
However, this is magic behaviour that is hard to maintain and
understand.

**Solution:**

Make using the desired LSP server explicit by disabling
`automatic_enable` of the LSP server, and then explicitly calling
`vim.lsp.enable("lua_ls")`.