test: speed up LSP rename tests by using emmylua_ls (#1031)
**Issue:**

In the project, there are e2e tests for testing file renaming between
Neovim, yazi, and an lsp server. They make sure that the lsp renaming
feature works with yazi.nvim in a real environment.

However, running the LSP rename tests takes 8+ seconds. It's a bit too
much, especially if we add more tests in the future.

**Solution:**

Use the modern EmmyLua LSP server
https://github.com/EmmyLuaLs/emmylua-analyzer-rust

This takes the time to run the tests down from 8+ seconds to 4+ seconds,
which is a bit better.

This LSP server is written in Rust (very important to mention 🙃), and
was recently taken into use in Neovim
https://github.com/neovim/neovim/pull/34323

Did I mention it's written in Rust?