refactor(tests): use integration plugins only for a minority of tests
**Issue:**

Currently in the integration-tests, 7 neovim plugins are installed so
that the integrations with them can be tested. However, most of the
tests do not use these plugins.

I'm worried that the presence of the plugins in the test neovim
directory (`integration-tests/test-environment/.repro/data/nvim`) will
affect the code that is tested at some point. Right now this is not an
issue, but it could be in the future.

**Solution:**

Make the integration-tests use two different neovim directories - one
for a minimal yazi.nvim setup with no integrations, and another one that
has all the integrations and the `emmylua_ls` LSP server installed.

I added this feature to tui-sandbox here
https://github.com/mikavilpas/tui-sandbox/pull/510, and tested it with a
real (small) project here
https://github.com/mikavilpas/tsugit.nvim/pull/249