test: add retry logic to `Yazi_reveal_path()` for flaky tests
This is an internal function that some tests use to reveal a file in
yazi.

Some tests sometimes try to reveal a path even though yazi is not ready
yet. It could be fixed by making each test wait for yazi to be ready,
but it's perhaps easier to add the retry logic to `Yazi_reveal_path()`.