perf: fix `Yazi toggle` always failing the first try
**Issue:**

When using `Yazi toggle` to open yazi and focus (reveal) the previously
hovered path, the first attempt to reveal the path often fails. This is
because the `reveal` action is executed in a `vim.in_fast_event()`.

The action is retried many times, and typically the next attempt
succeeds. However, it's slower than it could be.

**Solution:**

Use `vim.schedule_wrap` to ensure the action is executed in the main
thread, allowing it to run immediately without being delayed by
retrying.