feat: support installing plugins from deeply nested subdirectories
**Issue:**

Using the built-in lazy.nvim-based plugin management system, it's only
possible to install plugins from a subdirectory of the plugin's root
directory. This works fine for e.g. https://github.com/yazi-rs/plugins/
which provides all the plugins as exactly one level deep from the repo
root.

However, some plugins may be located in a deeply nested subdirectory
such as `deeply/nested/plugin.yazi`, and the current implementation puts
them in `~/.config/yazi/plugins/deeply/nested/plugin.yazi` which does
not work with yazi.

Docs for this feature:

- https://github.com/mikavilpas/yazi.nvim/blob/main/documentation/plugin-management.md

**Solution:**

Allow passing in a `name` option to the `build_plugin` and
`build_flavor` functions, which will be used as the name of the
symlinked plugin or flavor. This puts the plugin in
`~/.config/yazi/plugins/plugin.yazi` instead of
`~/.config/yazi/plugins/deeply/nested/plugin.yazi`.

This can be helpful in the future, in case we bundle plugins that
enhance the yazi.nvim <-> yazi integration.