ci: automerge dependabot updates when tests pass (#1128)
Dependabot is used to automatically create pull requests for testing
related dependencies. These do not affect the neovim functionality of
this plugin in any way, and users should not be concerned about these
updates.

This is an internal change done to reduce the maintenance burden of this
repository.

It will create more frequent commits to the repository, so users are
recommended to set the version manager to only update to the newest
releases, and not to the latest commit. I added an example of this to
the readme file.

- https://lazy.folke.io/spec/versioning
- https://github.com/echasnovski/mini.deps