fix: yazi window displays a nice name for other plugins to see (#1132)
**Issue:**

Plugins like https://github.com/nvim-lualine/lualine.nvim may use the
buffer name to display what buffer is the current one.

Currently the buffer name displays a lot of technical details, such as
the command that was used to start yazi.

**Solution:**

Display "yazi" instead. The issue below contains an example of how this
can be customized further, but this is a good default.

Closes https://github.com/mikavilpas/yazi.nvim/issues/1127