ci: install lua and emmylua-analyzer-rust with mise (#1183)
Mise provides a fast caching action for GitHub Actions. I want to try
moving many of the dependencies to be installed with mise, so that we
can avoid the delay of downloading and installing them on every run.

For example, here it took 1m38s to run
`apt-get install ripgrep fd-find fzf`:

- https://github.com/mikavilpas/yazi.nvim/actions/runs/17675113095/job/50235280065?pr=1183