refactor!: disable custom shell escaping as unneeded (#1217)
BREAKING CHANGE: removed `integrations.escape_path_implementation`
(function) and `future_features.new_shell_escaping` (boolean).

The custom shell escaping implementation is no longer needed, as we
found a better way to handle starting yazi in
https://github.com/mikavilpas/yazi.nvim/issues/1101 that should avoid
all kinds of shell related issues on all platforms.

It has been automatically enabled by default for almost two months
without any issues.

Nobody seems to be using this feature, and we haven't had any bug
reports about it since it was merged in
https://github.com/mikavilpas/yazi.nvim/pull/1102

I left the old implementation in as a precaution for users to optionally
toggle it back on, but I think it's safe to remove it completely.