feat: allow setting more window properties with a hook
Thanks to @and-rs for the idea!

This allows setting various options for the floating window, such as the
`row` and `col` where the window should be positioned on the screen.

Note that not everything can be set, currently the window still has to
be a floating window.

Supersedes https://github.com/mikavilpas/yazi.nvim/pull/1406

Co-authored-by: and-rs <deserd@protonmail.com>