test: adapt to :messages staying open on nightly neovim
**Issue:**

Neovim nightly blocks when `print()` messages are visible for some
reason. This makes a test fail, blocking updates and continuous test
runs that make sure the plugin always works with the latest versions of
everything.

**Solution:**

Printing the messages can be avoided as it's not checked in the test. We
already have a global variable that stores the events, so we can just
read that instead, avoiding the issue.