ci: avoid GITHUB_TOKEN issues with mise lua installation
**Issue:**

I suspect that mise installing lua is causing issues with the
GITHUB_TOKEN rate limits. It may not be using the token correctly,
leading to the build getting rate limited.

Example:
https://github.com/mikavilpas/yazi.nvim/actions/runs/19967417993/job/57262837513

**Solution:**

Lua does not have to be installed through mise if we copy the build
commands that install lua dependencies and just run them directly in the
build. Hopefully this will avoid the issues.